emailsystem/requirements.txt
blinker==1.9.0
boto3==1.36.11
botocore==1.36.11
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
serverless-wsgi==3.0.5
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
